test: added more e2e tests

+++ e2e/_run_all_tests.sh
@@ -3,6 +3,11 @@ source program_tests.sh
 source print_expr_tests.sh
 source println_expr_tests.sh
 source unary_expr_tests.sh
+source binary_add_expr_tests.sh
+source let_expr_tests.sh
+source if_expr_tests.sh
+source if_else_expr_tests.sh
+source cli_tests.sh
 
 count_total=0
 count_passed=0
@@ -36,11 +41,19 @@ if [ ! -f "$compiler_target_path" ]; then
   exit 1
 fi
 
-# Copy the compiler jar to the root directory
+# Copy the compiler jar into this directory
 cp $compiler_target_path .
 
-# Build the runtime and write the object file to the root directory
+# Build the runtime and write the object file into this directory
 gcc -c $runtime_source_path -o ./runtime.o
+
+VERBOSITY_QUIET='VERBOSITY_QUIET'
+VERBOSITY_LOUD='VERBOSITY_LOUD'
+if [[ "$1" == "--quiet" ]]; then
+  verbosity=VERBOSITY_QUIET
+else
+  verbosity=VERBOSITY_LOUD
+fi
 
 # At this point, we can now run all e2e tests.
 # Run tests, passing in the path to the compiler jar as the first argument
@@ -48,22 +61,19 @@ run_program_tests ./$compiler_file_name
 run_print_expr_tests ./$compiler_file_name
 run_println_expr_tests ./$compiler_file_name
 run_unary_expr_tests ./$compiler_file_name
+run_binary_add_expr_tests ./$compiler_file_name
+run_let_expr_tests ./$compiler_file_name
+run_if_expr_tests ./$compiler_file_name
+run_if_else_expr_tests ./$compiler_file_name
+run_cli_tests ./$compiler_file_name
 
 # Display statistics
 echo -e "Ran $count_total tests\n Passed: $count_passed\n Failed: $count_failed"
 
 # Cleanup
-if [ -f "a.out" ]; then
-  rm a.out
-fi
-
-if [ -f "runtime.o" ]; then
-  rm runtime.o
-fi
-
-if [ -f "$compiler_file_name" ]; then
-  rm $compiler_file_name
-fi
+if [ -f "a.out" ]; then rm a.out; fi
+if [ -f "runtime.o" ]; then rm runtime.o; fi
+if [ -f "$compiler_file_name" ]; then rm $compiler_file_name; fi
 
 # Exit
 if [ "$count_failed" -eq "0" ]; then
